requirements.txt
asgiref==3.11.1
Deprecated==1.3.1
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
lxml==6.0.2
Markdown==3.10.1
packaging==26.0
pikepdf==10.3.0
pillow==12.1.1
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
wrapt==2.1.1
